[feature][#3] chartApp 추가
차트 관련 화면영역인 chartApp을 추가했다.
추후 클라이언트 라우터를 만들어서 합칠 생각이다.
