+++ frontend/src/chartApp/chartApp.scss
@@ -1,5 +1,5 @@
-@import './src/utils/reset.scss';
-@import './src/component/header/header.scss';
+@import '../utils/reset.scss';
+@import '../component/header/header.scss';
 
 
 #app {
