+++ frontend/src/chartApp/index.html
@@ -1,0 +1,37 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+        content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Document</title>
+    <link rel="stylesheet" href="./chartApp/chartApp.css">
+    <style>
+        /* circle {
+            stroke-dasharray: 100;
+            stroke-dashoffset: 100;
+            animation: dash 1.3s linear alternate;
+        }
+
+        @keyframes dash {
+            0% {
+                stroke-dashoffset: 100;
+            }
+
+            100% {
+                stroke-dashoffset: 0;
+            }
+        } */
+    </style>
+</head>
+
+<body>
+    <div id="app"></div>
+
+
+
+</body>
+
+</html>
